third_party/WebKit/LayoutTests/css3/unicode-bidi-insolate-parse-expected.txt

Basic test for unicode-bidi: isolate parsing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testDiv.style.unicodeBidi is "-webkit-isolate"
PASS successfullyParsed is true

TEST COMPLETE
